requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=20.0,<21.0
openai>=1.0.0
supabase>=2.0.0
httpx>=0.26.0
numpy>=1.24.0
prometheus_client>=0.19.0
backoff>=2.2.1
circuitbreaker>=1.4.0
tenacity>=8.2.3
